+++ hosts/gaia/configuration.nix
@@ -1,0 +1,21 @@
+{ config, lib, ... }:
+
+{
+  imports = [ ./hardware-configuration.nix ];
+
+  boot.loader.systemd-boot.enable = true;
+  boot.loader.efi.canTouchEfiVariables = true;
+
+  networking.hostId = "8425e349";
+
+  services.openssh.openFirewall = lib.mkForce false;
+
+  services.tailscale.authKeyFile =
+    config.clan.core.vars.generators.tailscale.files.auth-key.path;
+
+  zramSwap.enable = true;
+  zramSwap.memoryPercent = 250;
+
+  # Check that this can be bumped before changing it
+  system.stateVersion = "25.11";
+}