+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Use `vars` to store Tailscale auth key
+- Added `gaia`
 - Use `networking.networkmanager.enableDefaultPlugins`
 - Disabled autoupdating in Spotify on macOS
 - Removed `firefox`, `remmina` and `spotify` overlays